[IMP] REF module and DEL Account dependency
