+++ l10n_cr_country_codes/__manifest__.py
@@ -1,5 +1,3 @@
-
-
 {
     'name': 'Códigos País para Facturación electrónica Costa Rica',
     'version': '15.0.1.0.0',
@@ -10,8 +8,7 @@
     'description': '''Códigos País para Facturación electrónica Costa Rica.''',
     'depends': [
         'base',
-        'account',
-        'product'
+        'contacts',
     ],
     'data': [
         'data/res_country_state.xml',
@@ -19,7 +16,9 @@
         'data/res.country.district.csv',
         'data/res.country.neighborhood.csv',
         'security/ir.model.access.csv',
-        'views/country_codes_views.xml',
+        'views/res_country_county_views.xml',
+        'views/res_country_district_views.xml',
+        'views/res_country_neighborhood_views.xml',
         'views/res_company_views.xml',
         'views/res_partner_views.xml',
     ],
