+++ l10n_cr_country_codes/models/__init__.py
@@ -1,3 +1,5 @@
-from . import country_codes
+from . import res_country_county
+from . import res_country_district
+from . import res_country_neighborhood
 from . import res_company
 from . import res_partner
