+++ l10n_cr_country_codes/models/res_partner.py
@@ -1,7 +1,5 @@
 # copyright  2018 Carlos Wong, Akurey SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-# Upgrade: Norlan Ruiz 2021
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import models, fields, api
