--- l10n_cr_country_codes/views/country_codes_views.xml
@@ -1,137 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    
-        <record id="view_res_country_county_form"
-                model="ir.ui.view">
-            <field name="name">res.country.county.form</field>
-            <field name="model">res.country.county</field>
-            <field name="arch"
-                   type="xml">
-                <form string="County">
-                    <sheet string="County">
-                        <group col="4">
-                            <field name="code"/>
-                            <field name="name"/>
-                            <field name="state_id"
-                                   options='{"no_open": True}'/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-        <record id="view_res_country_county_tree"
-                model="ir.ui.view">
-            <field name="name">res.country.county.tree</field>
-            <field name="model">res.country.county</field>
-            <field name="arch"
-                   type="xml">
-                <tree string="Counties">
-                    <field name="code"/>
-                    <field name="name"/>
-                    <field name="state_id"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="view_res_country_district_form"
-                model="ir.ui.view">
-            <field name="name">res.country.district.form</field>
-            <field name="model">res.country.district</field>
-            <field name="arch"
-                   type="xml">
-                <form string="District">
-                    <sheet string="District">
-                        <group col="4">
-                            <field name="code"/>
-                            <field name="name"/>
-                            <field name="county_id"
-                                   options='{"no_open": True}'/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-        <record id="view_res_country_district_tree"
-                model="ir.ui.view">
-            <field name="name">res.country.district.tree</field>
-            <field name="model">res.country.district</field>
-            <field name="arch"
-                   type="xml">
-                <tree string="Districts">
-                    <field name="code"/>
-                    <field name="name"/>
-                    <field name="county_id"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="view_res_country_neighborhood_form"
-                model="ir.ui.view">
-            <field name="name">res.country.neighborhood.form</field>
-            <field name="model">res.country.neighborhood</field>
-            <field name="arch"
-                   type="xml">
-                <form string="Neighborhood">
-                    <sheet string="Neighborhood">
-                        <group col="4">
-                            <field name="code"/>
-                            <field name="name"/>
-                            <field name="district_id"
-                                   options='{"no_open": True}'/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-        <record id="view_res_country_neighborhood_tree"
-                model="ir.ui.view">
-            <field name="name">res_country.neighborhood.tree</field>
-            <field name="model">res.country.neighborhood</field>
-            <field name="arch"
-                   type="xml">
-                <tree string="Neighborhoods">
-                    <field name="code"/>
-                    <field name="name"/>
-                    <field name="district_id"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="action_res_country_county"
-                model="ir.actions.act_window">
-            <field name="name">Counties</field>
-            <field name="res_model">res.country.county</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-        <record id="action_res_country_district"
-                model="ir.actions.act_window">
-            <field name="name">Districts</field>
-            <field name="res_model">res.country.district</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-        <record id="action_res_country_neighborhood"
-                model="ir.actions.act_window">
-            <field name="name">Neighborhoods</field>
-            <field name="res_model">res.country.neighborhood</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-        <menuitem id="menu_res_country_county"
-                  name="Counties"
-                  parent="account.account_management_menu"
-                  sequence="2"
-                  action="action_res_country_county"/>
-        <menuitem id="menu_res_country_district"
-                  name="Districts"
-                  parent="account.account_management_menu"
-                  sequence="2"
-                  action="action_res_country_district"/>
-        <menuitem id="menu_res_country_neighborhood"
-                  name="Neighborhoods"
-                  parent="account.account_management_menu"
-                  sequence="2"
-                  action="action_res_country_neighborhood"/>
-    
-</odoo>
